Refactor Notification structure: rename and move NotificationAPIController, update project references, and add request models for push notifications
