+++ common/service/service/Models/API/Requests/PushNotifyByLoginUserRequest.cs
@@ -1,0 +1,10 @@
+﻿namespace Nagiyu.Common.Service.Models.API.Requests
+{
+    public class PushNotifyByLoginUserRequest
+    {
+        /// <summary>
+        /// メッセージ
+        /// </summary>
+        public string Message { get; set; }
+    }
+}
