+++ common/web/Nagiyu.Common.Web.csproj
@@ -10,7 +10,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\common\service\auth\Nagiyu.Common.Auth.Service.csproj" />
+    <ProjectReference Include="..\service\auth\Nagiyu.Common.Auth.Service.csproj" />
+    <ProjectReference Include="..\service\service\Nagiyu.Common.Service.csproj" />
   </ItemGroup>
 
 </Project>
